fix warning and types

+++ src/xtd_drawing/src/xtd/drawing/image.cpp
@@ -51,7 +51,7 @@ void image::update_properties() {
     
   this->data_->pixel_format_ = static_cast<imaging::pixel_format>(native::image::pixel_format(this->data_->handle_));
   
-  size_t physical_width, physical_height;
+  int32_t physical_width, physical_height;
   native::image::physical_dimension(this->data_->handle_, physical_width, physical_height);
   this->data_->physical_dimension_ = drawing::size_f(static_cast<float>(physical_width), static_cast<float>(physical_height));
   
@@ -102,9 +102,9 @@ void image::update_properties() {
     case IFM_TGA: this->data_->raw_format_ = imaging::image_format::tga(); break;
   }
   
-  size_t width, height;
+  int32_t width, height;
   native::image::size(this->data_->handle_, width, height);
-  this->data_->size_ = drawing::size(static_cast<float>(width), static_cast<float>(height));
+  this->data_->size_ = drawing::size(width, height);
 
   this->data_->vertical_resolution_ = native::image::vertical_resolution(this->data_->handle_);
 }
